# Conflicts:
#	hudi-flink/src/main/java/org/apache/hudi/configuration/FlinkOptions.java
#	hudi-flink/src/main/java/org/apache/hudi/sink/StreamWriteFunction.java
#	hudi-flink/src/main/java/org/apache/hudi/sink/bulk/BulkInsertWriteFunction.java
#	hudi-flink/src/main/java/org/apache/hudi/sink/bulk/BulkInsertWriterHelper.java
#	hudi-flink/src/main/java/org/apache/hudi/sink/partitioner/BucketAssignFunction.java
#	hudi-flink/src/main/java/org/apache/hudi/sink/utils/Pipelines.java
#	hudi-flink/src/main/java/org/apache/hudi/source/StreamReadMonitoringFunction.java
#	hudi-flink/src/main/java/org/apache/hudi/streamer/FlinkStreamerConfig.java
#	hudi-flink/src/main/java/org/apache/hudi/table/HoodieTableSource.java
#	hudi-flink/src/main/java/org/apache/hudi/util/StreamerUtil.java

+++ hudi-client/hudi-client-common/src/main/java/org/apache/hudi/config/HoodieWriteConfig.java
@@ -1391,8 +1391,8 @@ public class HoodieWriteConfig extends HoodieConfig {
     return getInt(HoodieStorageConfig.LOGFILE_DATA_BLOCK_MAX_SIZE);
   }
 
-  public int getLogFileMaxSize() {
-    return getInt(HoodieStorageConfig.LOGFILE_MAX_SIZE);
+  public long getLogFileMaxSize() {
+    return getLong(HoodieStorageConfig.LOGFILE_MAX_SIZE);
   }
 
   public double getParquetCompressionRatio() {
